rearrange project
