+++ src/Data/ByteString/Base64/Internal.hs
@@ -17,29 +17,10 @@
 -- processes and tables.
 --
 module Data.ByteString.Base64.Internal
-( -- * Base64 encoding
-  encodeBase64_
+( encodeBase64_
 , encodeBase64Nopad_
-
-  -- * Base64 decoding
-, Padding(..)
 , decodeBase64_
 , decodeBase64Lenient_
-
-  -- * Decoding Tables
-  -- ** Standard
-, decodeB64Table
-  -- ** Base64-url
-, decodeB64UrlTable
-
-  -- * Encoding Tables
-  -- ** Standard
-, base64Table
-
-  -- ** Base64-url
-, base64UrlTable
-
-  -- * Validating Base64
 , validateBase64
 ) where
 
@@ -49,6 +30,7 @@ module Data.ByteString.Base64.Internal
 import Data.Bits
 import qualified Data.ByteString as BS
 import Data.ByteString.Base64.Internal.Tail
+import Data.ByteString.Base64.Internal.Types
 import Data.ByteString.Base64.Internal.Utils
 #if WORD_SIZE_IN_BITS == 32
 import Data.ByteString.Base64.Internal.W32.Encode
@@ -71,52 +53,6 @@ import GHC.Word
 
 import System.IO.Unsafe
 
--- -------------------------------------------------------------------------- --
--- Internal data
-
--- | Only the lookup table need be a foreignptr,
--- and then, only so that we can automate some touches to keep it alive
---
-data EncodingTable = EncodingTable
-  {-# UNPACK #-} !(Ptr Word8)
-  {-# UNPACK #-} !(ForeignPtr Word16)
-
--- | A type isomorphic to 'Bool' marking support for padding out bytestrings (@Pad),
--- or not (@Nopad@).
---
-data Padding
-    = Pad
-      -- ^ Do we pad out the bytestring?
-    | NoPad
-      -- ^ Do we not pad out the bytestring?
-    deriving Eq
-
--- | Pack an 'Addr#' into an encoding table of 'Word16's
---
-packTable :: Addr# -> EncodingTable
-packTable alphabet = etable
-  where
-    ix (I# n) = W8# (indexWord8OffAddr# alphabet n)
-
-    !etable =
-      let bs = concat
-            [ [ ix i, ix j ]
-            | !i <- [0..63]
-            , !j <- [0..63]
-            ]
-      in EncodingTable (Ptr alphabet) (writeNPlainForeignPtrBytes 8192 bs)
-
--- | Base64url encoding table
---
-base64UrlTable :: EncodingTable
-base64UrlTable = packTable "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"#
-{-# NOINLINE base64UrlTable #-}
-
--- | Base64 std encoding table
---
-base64Table :: EncodingTable
-base64Table = packTable "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/"#
-{-# NOINLINE base64Table #-}
 
 -- -------------------------------------------------------------------------- --
 -- Validating Base64
@@ -180,52 +116,6 @@ encodeBase64Nopad_ (EncodingTable !aptr !efp) (PS !sfp !soff !slen) =
 
 -- -------------------------------------------------------------------------- --
 -- Decoding Base64
-
--- | Non-URLsafe b64 decoding table (naive)
---
-decodeB64Table :: ForeignPtr Word8
-decodeB64Table = writeNPlainForeignPtrBytes @Word8 256
-      [ 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0x3e,0xff,0xff,0xff,0x3f
-      , 0x34,0x35,0x36,0x37,0x38,0x39,0x3a,0x3b,0x3c,0x3d,0xff,0xff,0xff,0x63,0xff,0xff
-      , 0xff,0x00,0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x09,0x0a,0x0b,0x0c,0x0d,0x0e
-      , 0x0f,0x10,0x11,0x12,0x13,0x14,0x15,0x16,0x17,0x18,0x19,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0x1a,0x1b,0x1c,0x1d,0x1e,0x1f,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28
-      , 0x29,0x2a,0x2b,0x2c,0x2d,0x2e,0x2f,0x30,0x31,0x32,0x33,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      ]
-{-# NOINLINE decodeB64Table #-}
-
--- | URLsafe b64 decoding table (naive)
---
-decodeB64UrlTable :: ForeignPtr Word8
-decodeB64UrlTable = writeNPlainForeignPtrBytes @Word8 256
-      [ 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0x3e,0xff,0xff
-      , 0x34,0x35,0x36,0x37,0x38,0x39,0x3a,0x3b,0x3c,0x3d,0xff,0xff,0xff,0x63,0xff,0xff
-      , 0xff,0x00,0x01,0x02,0x03,0x04,0x05,0x06,0x07,0x08,0x09,0x0a,0x0b,0x0c,0x0d,0x0e
-      , 0x0f,0x10,0x11,0x12,0x13,0x14,0x15,0x16,0x17,0x18,0x19,0xff,0xff,0xff,0xff,0x3f
-      , 0xff,0x1a,0x1b,0x1c,0x1d,0x1e,0x1f,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28
-      , 0x29,0x2a,0x2b,0x2c,0x2d,0x2e,0x2f,0x30,0x31,0x32,0x33,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      , 0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff
-      ]
-{-# NOINLINE decodeB64UrlTable #-}
 
 -- | The main decode function. Takes a padding flag, a decoding table, and
 -- the input value, producing either an error string on the left, or a
